requirements.txt
pandas>=2.0.0
pandasai>=2.0.0
mysql-connector-python>=8.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0
groq>=0.4.0
streamlit>=1.30.0
plotly>=5.18.0
pytest>=8.0.0
